Improve Makefile.
Replace CFLAGS_NO_WFLAGS with CFLAGS for some targets.
Change LIB to LIBB.
Add -Wno-old-style-declaration -Wno-unknown-warning-option to
  JudyMalloc.o recipe.
Separate JudyMalloc.c from LIBB1_SRCS and add separate
JudyMalloc.so target with
  -Wno-old-style-declaration -Wno-unknown-warning-option.
Use $^ in libb.a, libb1.a, libb.so and libb1.so recipes.
Add dependencies for libb.so and libb1.so.