Proof of concept for UA_PARALLEL_128.
We wanted to bump MALLOC_ALIGNMENT to 16 for 32-bit executables in
order to gain a type field bit in wRoot.
But doing so hurt memory usage pretty badly for low population
16-bit lists when using PSPLIT_PARALLEL and PARALLEL_128.
So we wrote this code to do a mostly 128-bit parallel psplit
search without requiring that the list ends be aligned.
Here are the basics:
Pad list to end of malloc buffer for haskey.
Do a 96-bit search on the 128-bit bucket chosen by psplit and return if
the key is found.
If the key is not found figure out if the last 32-bit word of the
bucket is valid and search that.
The proceed normally until we hit the end of the list in HASKEYF.
If we think we found the key we have to confirm that it is not in
a garbage word that doesn't belong to this malloc buffer.
The method depends on a priori knowledge about how malloc works.
The current code is activated with cnBitsPerWord=32 and
MALLOC_ALIGNMENT=16 to make sure front end of list is aligned and
UA_PARALLEL_128 to define NO_PARALLEL_128 avoid allocating
extra memory to align the list ends. We need to work on the ifdefs.