Try to get Check JudyHS testing to work.
Discovered an incompatibility between JudyHS and my JudyL.
JudyHS assumes it it free to use the JLAP_INVALID bit in
the root word of an array. But my JudyL uses that bit
and all of the others.

For now, we have to build Check with NO_TEST_HS if/when using
my JudyL.